requirements.txt
flask==2.3.2
python-dotenv==1.0.0
requests>=2.32.0
gunicorn>=22.0.0
pytest>=7.4.0
pytest-cov==4.1.0
cairosvg==2.7.1
Pillow==10.0.0 